Add requirejs support to karma-ng-html2js-preprocessor.
Users can enable requirejs by using the requirejs framework or adding the config
property of require to the karma config file.  The require config property may
contain a name to use when referring to the angular shim, by default it is
'angular'.

The README has been updated with this information and includes steps to take
to modify the test-main.js file.  The instructions supersede the the karma+
requirejs instructions for those using angular.